requirements.txt
# Core AI/ML Libraries
torch>=2.5.1
torchvision>=0.20.1
torchaudio>=2.5.1
transformers>=4.35.0
huggingface-hub>=0.17.0

# Web Interface
gradio>=3.50.0

# Video/Audio Processing
ffmpeg-python>=0.2.0
librosa>=0.10.0
soundfile>=0.12.0

# Image Processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
requests>=2.31.0
pyyaml>=6.0.1
tqdm>=4.66.0

# GPU/CUDA support
nvidia-cuda-runtime>=12.4.0  # This will be installed via conda

# Development dependencies (install with pip install -r requirements-dev.txt)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.0.0
# mypy>=1.5.0